Escape additional special characters (#3)
* Add ^ and $ to list of escaped characters

A `scanf` format specification such as `'^(%f)$'` should match a literal `^`, a floating point value and a literal `$`. However, these two characters are not escaped, so instead the resulting regex matches a string that consists only of a floating point value.

* Add | to special characters

According to the documentation of `re`, | is also a special symbol that should be escaped to make literal matches without manual escaping possible.